Small Changes to Schemas
This adds primary keys to the Cassandra schema in order to partition
the schema by city and then room_id.

For the PostgreSQL schema, it updates a few of the names and the order
so that it will be loaded into the DB. Also it adds in the serial key
word for the PKs.